4.4 loggers added 2nd properties with new port controller for port new profile test liquebase off

+++ src/main/java/ru/hogwarts/school/service/StudentServiceImpl.java
@@ -2,6 +2,7 @@ package ru.hogwarts.school.service;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import ru.hogwarts.school.exception.StudentNotFoundException;
 import ru.hogwarts.school.model.Student;
 import org.springframework.stereotype.Service;
 import ru.hogwarts.school.repo.CountRepo;
@@ -9,6 +10,7 @@ import ru.hogwarts.school.repo.StudentRepo;
 
 import java.util.Collection;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class StudentServiceImpl implements StudentService {
@@ -32,7 +34,7 @@ public class StudentServiceImpl implements StudentService {
     @Override
     public Student findStudent(long id) {
         logger.info("Was invoked method find student by id {}",id);
-        Student findStud = this.studentRepo.findById(id).get();
+        Student findStud = this.studentRepo.findById(id).orElseThrow(StudentNotFoundException::new);
         logger.info("Student with id {} is {}",id,findStud);
         return findStud;
     }
@@ -40,6 +42,10 @@ public class StudentServiceImpl implements StudentService {
     @Override
     public Student editStudent(long id, Student student) {
         logger.info("Was invoked method edit student by id {}",id);
+//        if (!this.studentRepo.findAll().contains(id)) {
+//            logger.error("Student {} not found and cant be edit",id);
+//            throw new StudentNotFoundException();
+//        }
         Student newStud = this.studentRepo.save(student);
         logger.info("Student with id {} was edit",id);
         return newStud;
@@ -48,6 +54,10 @@ public class StudentServiceImpl implements StudentService {
     @Override
     public void deleteStudent(long id) {
         logger.info("Was invoked method delete student by id");
+        if (!this.studentRepo.findAll().contains(id)) {
+            logger.error("Student with id {} not found",id);
+            throw new StudentNotFoundException();
+        }
         this.studentRepo.deleteById(id);
         logger.info("student with id {} was deleted",id);
     }
@@ -56,6 +66,10 @@ public class StudentServiceImpl implements StudentService {
     public Collection<Student> findByAge(int age) {
         logger.info("Was invoked method find student by age");
         Collection<Student> byAge = studentRepo.findByAge(age);
+        if (byAge.isEmpty()) {
+            logger.error("Student with age {} not found",age);
+            throw new StudentNotFoundException();
+        }
         logger.info("Student by age {} is {}",age,byAge);
         return byAge;
     }
@@ -64,6 +78,10 @@ public class StudentServiceImpl implements StudentService {
     public Collection<Student> getAllStud() {
         logger.info("Was invoked method for all students");
         Collection<Student> all = this.studentRepo.findAll();
+        if (all.isEmpty()) {
+            logger.error("There are no any students");
+            throw new StudentNotFoundException();
+        }
         logger.info("All students: {}",all);
         return all;
     }
@@ -72,6 +90,10 @@ public class StudentServiceImpl implements StudentService {
     public Collection<Student> findAllByAgeBetween(int from, int to) {
         logger.info("Was invoked method to find all students between ages");
         Collection<Student> findAll = this.studentRepo.findAllByAgeBetween(from, to);
+        if (findAll.isEmpty()) {
+            logger.error("There are no any students between {} and {}",from,to);
+            throw new StudentNotFoundException();
+        }
         logger.info("Students between age {} and {} are {}",from,to,findAll);
         return findAll;
     }
@@ -79,6 +101,10 @@ public class StudentServiceImpl implements StudentService {
     public Integer getStudentsCount() {
         logger.info("Was invoked method to get Number of students");
         Integer count = countRepo.getAllByIdIsNotNull();
+        if (this.studentRepo.findAll().isEmpty()) { //to check
+            logger.error("We have 0 Students");
+            throw new StudentNotFoundException();
+        }
         logger.info("There are {} students",count);
         return count;
     }
@@ -86,6 +112,10 @@ public class StudentServiceImpl implements StudentService {
     public Double getAverageAge() {
         logger.info("Was invoked method of average Age");
         Double avg = countRepo.getAverageAge();
+        if (this.studentRepo.findAll().isEmpty()) { //to check
+            logger.error("We have 0 Students");
+            throw new StudentNotFoundException();
+        }
         logger.info("Average age of students is {}",avg);
         return avg;
     }
@@ -93,6 +123,10 @@ public class StudentServiceImpl implements StudentService {
     public List<Student> getStudGreaterIdDesc5() {
         logger.info("Was invoked method of 5 last stud by id");
         List<Student> studentList = countRepo.getStudentByAgeIsGreaterThanOrderById();
+        if (this.studentRepo.findAll().isEmpty()) { //to check
+            logger.error("We have 0 Students");
+            throw new StudentNotFoundException();
+        }
         logger.info("Last 5 students by id : {} ",studentList);
         return studentList;
     }
